requirements.txt
pyudev==0.24.4
PyYAML==6.0.3
evdev==1.9.1
